backend/requirements.txt
fastapi>=0.75.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4
sqlmodel>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
pydantic[email]>=1.8.1
python-dotenv>=0.21.0

# Optional / dev dependencies (recommended for testing, migrations)
alembic>=1.8.0   # 数据迁移
pytest>=7.0.0
httpx>=0.23.0
pytest-asyncio>=0.20.0
# requests>=2.28.0
